No item View added with a veryyyy cool animation button

+++ ToDoList/ToDoList/Views/ListView.swift
@@ -16,35 +16,42 @@ struct ListView: View {
     @State private var tapPosition: CGPoint = .zero
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ZStack {
-                    ListRowView(item: item)
-                        .onTapGesture { location in
-                            let globalX = location.x
-                            let globalY = location.y
-                            tapPosition = CGPoint(x: globalX, y: globalY)
-                            counterForConfetti += 1
-                            withAnimation(.linear) {
-                                listViewModel.updateItem(item: item)
-                                
-                            }
+        ZStack {
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ZStack {
+                            ListRowView(item: item)
+                                .onTapGesture { location in
+                                    let globalX = location.x
+                                    let globalY = location.y
+                                    tapPosition = CGPoint(x: globalX, y: globalY)
+                                    counterForConfetti += 1
+                                    withAnimation(.linear) {
+                                        listViewModel.updateItem(item: item)
+                                    }
+                                }
+                            ConfettiCannon(counter: $counterForConfetti,
+                                           num: 15,
+                                           colors: [.red, .blue, .green, .yellow, .purple],
+                                           rainHeight: 400,
+                                           radius: 150)
+                                .position(tapPosition)
                         }
-                    ConfettiCannon(counter: $counterForConfetti,
-                                   num: 15,
-                                   colors: [.red, .blue, .green, .yellow, .purple],
-                                   rainHeight: 400,
-                                   radius: 150)
-                        .position(tapPosition)
+                        
+                        
+                    }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                    
                 }
-                
-                
+                .listStyle(PlainListStyle())
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
-            
         }
-        .listStyle(PlainListStyle())
+        
         .navigationTitle("To Do List 📝")
         .navigationBarItems(
             leading: EditButton(),
